moved main menu functionality to own package
